fix progress bug and change structure

+++ Pod/Classes/UI/CommonUI/CKDownloadProgressViewDelegate.h
@@ -12,6 +12,6 @@
 
 @required
 
-- (void)setProgress:(float)progress animated:(BOOL)animated ;
+- (void)setProgress:(CGFloat)progress animated:(BOOL)animated ;
 
 @end
